(feature): refactor tests and components
- refactor all files using SideMenuComponent to use the new NavBarComponent
