+++ src/__test__/AssetsSubCategoriesComponent.test.js
@@ -26,14 +26,6 @@ describe('Renders <AssetsSubCategoriesComponent /> correctly', () => {
     expect(wrapper.find('LoaderComponent').length).toBe(1);
   });
 
-  it('renders TableRowComponent component if isLoading is false', () => {
-    props.isLoading = false;
-    wrapper = shallow(<AssetSubCategoriesComponent
-      {...props}
-    />);
-    expect(wrapper.find('SideMenuComponent').length).toBe(1);
-  });
-
   it('renders Pagination component', () => {
     props.isLoading = false;
     wrapper = shallow(<AssetSubCategoriesComponent
