--- src/_css/DashboardComponent.css
@@ -1,13 +1,0 @@
-@media screen and (min-width: 960px) {
-    .chartArea {
-        display: flex;
-        flex-direction: row;
-        justify-content: space-around;
-    }
-}
-
-@media screen and (max-width: 960px) {
-    .fusioncharts-container {
-        width: 82% !important;
-    }
-}
