--- src/_css/NavBarComponent.css
@@ -1,16 +1,0 @@
-.navbar {
-  background: rgb(248, 249, 251);
-}
-
-.ui.secondary.menu > a {
-  font-size: 1.5rem;
-}
-
-.ui.menu .item > i.icon {
-  font-size: 1.5rem;
-}
-
-.user-avatar {
-  padding-left: 0 !important;
-  padding-right: 1rem !important;
-}
