+++ src/components/AssetTypesComponent.jsx
@@ -7,7 +7,7 @@ import _ from 'lodash';
 
 import TableRowComponent from './TableRowComponent';
 import AssetTypesAction from './AssetTypesAction';
-import SideMenuComponent from '../_components/SideMenuComponent';
+import NavbarComponent from './NavBarComponent';
 import LoaderComponent from '../components/LoaderComponent';
 import '../_css/AssetTypesComponent.css';
 
@@ -32,24 +32,24 @@ export class AssetTypesComponent extends React.Component {
   render() {
     if (this.props.isLoading) {
       return (
-        <SideMenuComponent>
+        <NavbarComponent>
           <LoaderComponent size="large" dimmerStyle={{ height: '90vh' }} />
-        </SideMenuComponent>
+        </NavbarComponent>
       );
     }
     if (!this.props.isLoading && _.isEmpty(this.props.assetTypes)) {
       return (
-        <SideMenuComponent>
+        <NavbarComponent>
           <Container>
             <h1>
               No Asset Types Found
             </h1>
           </Container>
-        </SideMenuComponent>
+        </NavbarComponent>
       );
     }
     return (
-      <SideMenuComponent>
+      <NavbarComponent>
         <Container>
           <Header className="landing-heading" content="Asset Types" />
           <Table celled>
@@ -93,7 +93,7 @@ export class AssetTypesComponent extends React.Component {
             </Table.Footer>
           </Table>
         </Container>
-      </SideMenuComponent>
+      </NavbarComponent>
     );
   }
 }
